file: 기존 BankManager 파일을 LinkedList, Queue 파일로 분리
